+++ app/components/omnibox/controllers/point-controller.js
@@ -57,6 +57,7 @@ angular.module('lizard-nxt')
 
             $scope.box.content[response.layerGroupSlug].layers[response.layerSlug].aggWindow = aggWindow;
           }
+          $scope.box.minimizeCards();
         });
       });
       // Draw feedback when all promises resolved
@@ -206,20 +207,5 @@ angular.module('lizard-nxt')
       ClickFeedbackService.emptyClickLayer($scope.mapState);
     });
 
-    /**
-     * @function
-     * @memberOf app.pointCtrl
-     * @description Toggling the view on the table for structure attributes;
-     *              Either show the first 3 attributes, OR show all of them
-     */
-    $scope.box.toggleFullTable = function () {
-
-      $scope.box.showFullTable = !$scope.box.showFullTable;
-
-      d3.selectAll('tr.attr-row')
-        .classed('hidden', function (_, i) {
-          return i > 2 && !$scope.box.showFullTable;
-        });
-    };
   }
 ]);