+++ app/components/omnibox/directives/omnibox-directives.js
@@ -40,6 +40,46 @@ angular.module("omnibox", ["templates-main"])
         scope.box.showCards = scope.box.type !== 'empty';
       };
 
+      /**
+       *  This resets the detailed card model
+       */
+      var resetDetailCards = function () {
+        scope.box.fullDetailCards = {};
+        angular.forEach(Object.keys(scope.box.content), function (key) {
+          scope.box.fullDetailCards[key] = true;
+          if (key === 'waterchain') {
+            scope.box.fullDetailCards.timeseries = true;
+          }
+        });
+      };
+
+      /**
+       * Resizes if cards, navbar and timeline are larger
+       * then the window size.
+       */
+      scope.box.minimizeCards = function () {
+        // height of search and nav combined
+        var searchNav = $('#searchboxinput').offset().top + $('#searchboxinput').height();
+        var cardsTooHigh = $('#cards').height() + searchNav +
+          $('#timeline').height() > $('body').height();
+        // jquery is good at this stuff alternative version would be:
+        // document.querySelector('#cards').clientHeight etc...
+        if (cardsTooHigh) {
+          angular.forEach(Object.keys(scope.box.fullDetailCards),
+            function (layer) {
+              scope.box.fullDetailCards[layer] = false;
+            }
+          );
+        }
+      };
+
+      window.addEventListener('resize', scope.box.minimizeCards);
+
+      scope.$watch('box.content', function () {
+        resetDetailCards();
+        scope.box.minimizeCards();
+      });
+
       scope.$watch('box.type', function (n, o) {
         if (n === o) { return true; }
         finalizeTemplateRendering();